05-Professional-Section/Day-97 - Professional Portfolio Project - [Web Development]/An Online Shop/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
stripe==5.5.0
python-dotenv==1.0.0
requests==2.31.0
